Added files from other members writeup
Left me with a moderate amount of work to do...
They didn't do a satisfactory job on this and the deal was I make the program and they do the report, and here I am having to now touch up the report. From what I read here it looks like they almost took 1 for 1 what I made in my README here.